requirements.txt
openai
datasets
faiss-cpu
numpy
pandas
sentence-transformers
torch
transformers
networkx
pandas
tqdm
python-dotenv
spacy
tensorboard>=1.15.0
ujson
streamlit
argparse
tiktoken